Change build timeout slicer to use XStream
The build timeout slicer has introduced a more object oriented
configuration with timeout strategies, which is difficult
for the slicer to follow in a static way. So instead the slicer
now leverages the XStream serializer, which should match any updates
to the build timeout plugin, at the expense of a more verbose
slice language.

Also, the option to slice more narrowly, such as BuildTimeoutSlicer2
provided is no longer trivially possible.  Something like XPath might
be useful to make that happen in the future though.