[JENKINS-48738] Fix JDK configuration NPE
https://github.com/jenkinsci/configurationslicing-plugin/pull/22/commits/a15d15ae0e3a666d30acf88f2156875d867e977e

Confirmed with interactive testing that the null pointer exception is
resolved and that the configuration slicing plugin can change the value
of the JDK that is defined on a Freestyle job.  I changed between jdk-8
and jdk-20.

The plugin cannot reset the definition of the JDK to the default because
that causes a null pointer exception in AbstractProject as reported by
the original bug report https://issues.jenkins.io/browse/JENKINS-48738

Avoid the null pointer exception by ignoring the attempt to reset the
JDK value to default.

Thanks to @Evildethow and @ahertier for implementation