Reduce Jenkinsfile diffs to archetype (#222)
Use the Jenkins plugin archetype as much as possible as we can so that
the differences to the archetype are clear and intentionally used.

Tests with Java 21 on Linux and Java 17 on Windows.  Accepts that tests
on Java 11 are redundant since we've not detected relevant issues that
are specific to Java 11.

Confirmed that tests pass on Linux with Java 17 and Java 21.

Rely on ci.jenkins.io to test Windows.

Includes forkCount setting because it improves test performance on
ci.jenkins.io and reduces cost on ci.jenkins.io.